17569 add images and get images

+++ src/main/java/com/softjourn/vending/dao/ImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface ImageRepository extends CrudRepository<Image,Long> {
+public interface ImageRepository extends CrudRepository<Image,Integer> {
 
     List<Image> findByProductId(Integer id);
 }
